simplify frontend
simplify frontend to reflect client ID only, remove modal.
